move Customer, Employee & Language query builders out of Query/Builder/ to Query/ namespace
