+++ src/PrestaShopBundle/Resources/config/services/core/grid/doctrine_query_builder.yml
@@ -43,7 +43,7 @@ services:
         public: true
 
     prestashop.core.grid.query_builder.employee:
-        class: 'PrestaShop\PrestaShop\Core\Grid\Query\Builder\EmployeeQueryBuilder'
+        class: 'PrestaShop\PrestaShop\Core\Grid\Query\EmployeeQueryBuilder'
         parent: 'prestashop.core.grid.abstract_query_builder'
         arguments:
             - '@prestashop.core.query.doctrine_search_criteria_applicator'
@@ -61,7 +61,7 @@ services:
             - "@=service('prestashop.adapter.shop.context').getContextListShopID()"
 
     prestashop.core.grid.query_builder.customer:
-        class: 'PrestaShop\PrestaShop\Core\Grid\Query\Builder\CustomerQueryBuilder'
+        class: 'PrestaShop\PrestaShop\Core\Grid\Query\CustomerQueryBuilder'
         parent: 'prestashop.core.grid.abstract_query_builder'
         arguments:
             - '@prestashop.core.query.doctrine_search_criteria_applicator'
@@ -70,7 +70,7 @@ services:
         public: true
 
     prestashop.core.grid.quer_.builder.language:
-        class: 'PrestaShop\PrestaShop\Core\Grid\Query\Builder\LanguageQueryBuilder'
+        class: 'PrestaShop\PrestaShop\Core\Grid\Query\LanguageQueryBuilder'
         parent: 'prestashop.core.grid.abstract_query_builder'
         arguments:
             - '@prestashop.core.query.doctrine_search_criteria_applicator'
